% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buque.R
\name{buque}
\alias{buque}
\title{buque}
\usage{
buque(
  digitado,
  filename = "Y:/Dinarama/Codigos_buque.csv",
  categ = TRUE,
  pareja = FALSE
)
}
\arguments{
\item{digitado}{el nombre del buque digitado en los partes de pesca}

\item{filename}{la ubicación del archivo base de datos por default es "Y:/Dinarama/Codigos_buque.csv"}

\item{categ}{queres incluir la categorìa del buque? =T}

\item{pareja}{es el barco principal? =F o la pareja =T}
}
\value{
vector buque con los nombres de los buques con categoria o sin ella
}
\description{
Accede la base de datos de buques y corrige los nombres mal digitados, proporciona categoria del buque
Base datos buques: columnas digitado,	buque,	buque_id,	categ,	potHP,	eslora
Incluye mensaje de error para buques no encontrados
}
\examples{
digitado<- c("ANTARES", "ANTARES 1", "ANTARES I")
buque(digitado, categ=TRUE)

}
