% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy.R
\name{copy}
\alias{copy}
\title{copy}
\usage{
copy(x)
}
\arguments{
\item{x}{la tabla, dataframe o matriz que queremos copiar}
}
\value{
una tabla separada por ";" y con decimales "." que queda en el portapapeles
}
\description{
funcion barata que copia una tabla al portapapeles y asi lo podes pegar en Excel sin comerte el coco
}
\examples{
copy(mvd)
#anda y pegalo en otro lado, besis
}
