% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_jabba.R
\name{cr_jabba}
\alias{cr_jabba}
\title{cr_jabba}
\usage{
cr_jabba(modelo)
}
\arguments{
\item{modelo}{= un modelo JABBA}
}
\value{
captura de reemplazo de las medias (mu) y del intervalo de credibilidad superior (uci) e inferior (lci)
}
\description{
Calcula la captura de reemplazo para un modelo JABBA con sus intervalos de credibilidad
}
