% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cuadzcp}
\alias{cuadzcp}
\title{subcuadrantes de pesca en la ZCP}
\format{
Un data frame con 3 columnas y pila de filas:
\describe{
\item{id}{el número de cuadrante y la letra que identifica subcuadrante}
\item{geometry}{la geometrìa de cada sub cuad}
\item{m2}{el area en m2 de cada subcuadrante}
...
}
}
\usage{
cuadzcp
}
\description{
Una base de datos que contiene las geometrias de los subcuadrantes incluidos en la ZCP
asì como su nombre y el area total
}
\keyword{datasets}
