% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_posicion.R
\name{d_posicion}
\alias{d_posicion}
\title{d_posicion}
\usage{
d_posicion(cuad, subcuad)
}
\arguments{
\item{cuad}{columna con cuadrantes}

\item{subcuad}{la columna con los subcuadrante}
}
\value{
dos columnas latR y longR (Recalculada)
}
\description{
devuelve la posicion geografica recalculada a partir de el cuadrante y subcuadrante
asigna todos las posiciones al centro del cuadrante facilitando el uso de geom_tile para plotear
}
\examples{
cuad<-c(345,365,363,361)
subcuad<-c("c","d", "b","a" )
d_posicion(cuad=cuad, subcuad=subcuad)

}
