% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emis_max_t.R
\name{emis_max_t}
\alias{emis_max_t}
\title{tiempo maximo sin emisiones}
\usage{
emis_max_t(datos, barco, unidad = "hours")
}
\arguments{
\item{datos}{la planilla de info vms con el tiempo en formato POSIXct usar tiempo_vms}

\item{barco}{el nombre del buque}

\item{unidad}{unidad de tiempo en ingles ("hours", "minutes")}
}
\value{
el maximo tiempo sin emisiones entre la primera y la ultima emision
}
\description{
calcula el tiempo maximo sin emisiones para un barco (en horas) desde información vms
}
