% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_check.R
\name{geo_check}
\alias{geo_check}
\title{geo_check}
\usage{
geo_check(lat, long, poly = "tierra")
}
\arguments{
\item{lat}{latitud del lance}

\item{long}{longitud del lance}

\item{poly}{el poligono con el que intersectamos los lances}
}
\value{
un vector binario indicando 1=sobre tierra y 0 sobre agua y alternativamente un plot
}
\description{
Un metodo rapido para chequear lances de pesca que caen en tierra o puertos
el default "tierra" usa un mapa de Sudamerica
si usamos "mvd" se aplica un poligono que abarca el puerto de Montevideo
si usamos "lpl" se aplica un poligono que abarca el puerto de La Paloma
}
\examples{
\dontrun{
long<-c(-55,-55,-55,-55,-55, -56.204417, -54.143567)
lat<-c(-37,-36,-35,-34,-33, -34.897430, -34.652272)
geo_check(lat=lat, long=long, poly="tierra")}
}
