% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lpl}
\alias{lpl}
\title{lpl}
\format{
Un data frame con 2 columnas y 12 filas:
\describe{
\item{long}{longitud geografica}
\item{lat}{latitud geografica}
...
}
}
\usage{
lpl
}
\description{
Una serie de 12 puntos georreferenciados que delimitan el puerto de La Paloma
(el primer y ultimo punto se repiten)
}
\keyword{datasets}
