% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mvd}
\alias{mvd}
\title{mvd}
\format{
Un data frame con 2 columnas y 9 filas:
\describe{
\item{long}{longitud geografica}
\item{lat}{latitud geografica}
...
}
}
\usage{
mvd
}
\description{
Una serie de 9 puntos georreferenciados que delimitan el puerto de Montevideo
(el primer y ultimo punto se repiten)
}
\keyword{datasets}
