% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomspp.R
\name{nomspp}
\alias{nomspp}
\title{nomspp}
\usage{
nomspp(
  digitado,
  filename = "Y:/Dinarama/Codigos_nom_spp.csv",
  spp = F,
  familia = F,
  orden = F
)
}
\arguments{
\item{digitado}{el nombre de especie digitado en el parte de pesca}

\item{filename}{= "Y:/Dinarama/Codigos_nom_spp.csv"}

\item{spp}{queres la especie? =T}

\item{familia}{queres la familia? =T}

\item{orden}{queres el orden? =T}
}
\value{
vector nom_spp con los nombres comunes acordados mas info taxonomica
}
\description{
accede a la base de datos de especies y brinda un nombre común acordado, con opcion a obtener especie, familia y orden
Utilizar con cuidado!! muchas veces la especie cambia con la categorìa o zona de la operativa de pesca, ante la duda chequealo con Ernesto!
Ademas en caso de trabajar con descarte usar filename="Y:/Dinarama/Codigos_descarte.csv"
Base datos especie: columnas digitado,	nom_spp,	spp,	familia,	orden.
Incluye mensaje de error para especies no encontradas
}
\examples{
lista_especies<- c("MERLUZA CARIOCA", "MERLUZA CCC", "MERLUZA CHICA",
 "MERLUZA COLA", "MERLUZA de COLA","MERLUZA DE COLA", "MERLUZA de COLA   HOKY ",
 "MERLUZA.de.COLA...HOKY.", "MERLUZA.CARIOCA")
 nomspp(lista_especies)
 nomspp(lista_especies, spp=TRUE)
}
