% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posicion.R
\name{posicion}
\alias{posicion}
\title{posicion}
\usage{
posicion(lat, long, hex = T, cuad = T, subcuad = F)
}
\arguments{
\item{lat}{latitud}

\item{long}{longitud}

\item{hex}{si las posiciones estan en formato hexadecimal dinara (ej. 3526) hex=T, si estan en decimales hex=F}

\item{cuad}{queres el cuadrante? =T}

\item{subcuad}{queres el subcuadrante? =T}
}
\value{
una tabla con lat, long (decimales), cuad y subcuad dependiendo de las opciones
}
\description{
transforma posiciones geograficas hexadecimal a decimal, asigna cuadrante y subcuadrante
}
\examples{
lat<-c(3526,3510,3450)
long<-c(5430, 5200, 5210)
posicion(lat, long)
}
