% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof.R
\name{prof}
\alias{prof}
\title{prof}
\usage{
prof(lat, long)
}
\arguments{
\item{lat}{latitud del lance}

\item{long}{longitud del lance}
}
\value{
un vector indicando la profundidad en metros
}
\description{
devuelve la profundidad de un set de coordenadas geográficas basado en data:
class      : RasterLayer;
dimensions : 185, 185, 34225  (nrow, ncol, ncell);
crs        : +proj=longlat +datum=WGS84 +no_defs;
values     : 16.276, 17.65286  (min km2, max km2).
}
\examples{
long<-c(-55,-55,-55,-55,-55, -56.204417, -54.143567)
lat<-c(-37,-36,-35,-34,-33, -34.897430, -34.652272)
prof(lat=lat, long=long)
}
