% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiempo.R
\name{tiempo}
\alias{tiempo}
\title{tiempo}
\usage{
tiempo(input, h_cala, h_vira, format = "\%d/\%m/\%Y")
}
\arguments{
\item{input}{la fecha del lance}

\item{h_cala}{la hora de cala del lance formato texto  ej. "11:30"}

\item{h_vira}{la hora de vira del lance formato texto  ej. "0:15"}

\item{format}{"\%d/\%m/\%Y"...el formato de la fecha input}
}
\value{
tabla con 3 columnas t_cala y t_vira (POSIXct) y h_pesca (horas de pesca)
}
\description{
construye tiempo de cala y vira en formato POSIXct ("\%Y-\%m-\%d \%H:\%M") y calcula las horas de pesca (h_pesca)
corrige la fecha de vira si se pasan las 00:00
}
\examples{
cala<- c("11:15", "10:20", "19:30")
vira<- c("15:37", "19:00", "06:00")
fecha<- c("12/10/2021","13/10/2021", "13/10/2021" )
tiempo(fecha, cala, vira)
}
