% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiempo_vms.R
\name{tiempo_vms}
\alias{tiempo_vms}
\title{tiempo_vms}
\usage{
tiempo_vms(fecha, hora, format = "\%d/\%m/\%Y")
}
\arguments{
\item{fecha}{la fecha, como sale del vms}

\item{hora}{la hora, como sale del vms}

\item{format}{="\%d/\%m/\%Y" es el default, se puede modificar}
}
\value{
un vector con la fecha POSIXct ("\%Y-\%m-\%d \%H:\%M")
}
\description{
construye tiempo en formato POSIXct ("\%Y-\%m-\%d \%H:\%M") para datos de vms
}
\examples{
fecha<- c("25/1/2022", "25/1/2022", "25/1/2022", "25/1/2022", "25/1/2022", "25/1/2022")
hora<- c("0:10", "1:10", "2:12", "3:12", "4:12","5:12")
tiempo_vms(fecha, hora)
}
