% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vel_arr_vms.R
\name{vel_arr_vms}
\alias{vel_arr_vms}
\title{vel_arr_vms}
\usage{
vel_arr_vms(vel, vmax = 4.5, vmin = 3.5)
}
\arguments{
\item{vel}{velocidad vms en mn/h(nudos)}

\item{vmax}{velocidad maxima considerada como arrastre < o =4.5}

\item{vmin}{velocidad minima considerada como arrastre > o =3.5}
}
\value{
un vector binario donde los 1's indican velocidad de arrastre
}
\description{
indica si la velocidad corresponde a velocidad de arrastre vms
}
\examples{
vel<-c(2,3,4,5,6,4,3.8)
vel_arr_vms(vel)
}
