% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z.R
\docType{data}
\name{z}
\alias{z}
\title{First five letters of the alphabet}
\format{
A data frame with 5 rows and 2 variables:
\describe{
\item{id}{integer from 1 to 5}
\item{names}{letter, from A to E}
...
}
}
\usage{
z
}
\description{
A dataset containing first five letters of English alphabet
}
\keyword{datasets}
